# simplify LetterNode, improve cache behaviour