EDIT: I just saw the Unicode PR, which makes this one superfluous. Sorry for taking your time 

I made some changes to LetterNode almost 2 years ago that reduced memory usage and improved performance. There were a few niggles I had about it however: every node had a vector, but only 1 actually used it and we were using indexes in the main algorithm and storing indexes in the data structure at the same time. This PR is intended to fix that. 

The main change is that most of the data is no longer stored: the bool storing case of letters was unused, the indices were already implicit in the indexes used in Candidate and so could be removed as well. In the end, all that is stored is one array of offsets per letter, all in one vector which should improve cache locality.

I've used a similar setup for benchmarking as described [here](https://gist.github.com/dgel/0ee4709015288ead82c14e2c079710bd) for the previous PR. some results (using 1000 samples):

shorthands:

cs: case-sensitive
cis: case-insensitive
lc: lowercase
mc: mixed case

## Speed

 variant | construction | match cis, lc | match cs, lc | match cis, mc | match cs, mc 
---------|--------------|---------------|--------------|---------------|--------------
 original | 11.1 us | 416 ns | 358 ns | 409 ns | 431 ns 
 modified | 7.9 us | 201 ns | 337 ns | 220 ns | 324 ns

## memory usage

variant | 100000 nodes of 4 chars | 10000 nodes of 78 chars
-------|-------------------------|------------------------
original | 220MB | 422MB
modified | 200MB | 383MB

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/valloric/ycmd/935)
<!-- Reviewable:end -->