@puremourning 2018-02-20
Wow! Thanks for making a PR!


> EDIT: I just saw the Unicode PR, which makes this one superfluous. 

I haven't looked at the detail of this, but I wonder if there are aspects of this PR we can learn from/use in the Unicode PR, or could do as a follow up? I see you've added some comments on that PR, which are super valuable, so thanks for that too!