@dgel 2018-02-21
You're welcome, I've been happily using YouCompleteMe for several years now and wanted to contribute a small thing back =) I don't think any of this is directly applicable to the Unicode PR, since it just does a direct search over the candidates. I think that is the right way to go, so not much room to improve there.