@Valloric 2018-02-27
Thanks for your amazing PR! :)

Like @puremourning, the unicode PR is making this a bit complicated... had this PR come in before that one, it would have likely been merged.

>  I don't think any of this is directly applicable to the Unicode PR, since it just does a direct search over the candidates. I think that is the right way to go, so not much room to improve there.

It's unfortunate that we can't get some knowledge from this PR over to that one... Should we close this PR then if there's not much that can be merged given the unicode PR?